% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagExtraction_Function.R
\name{CellTagExtraction}
\alias{CellTagExtraction}
\title{CellTag Extraction Function}
\usage{
CellTagExtraction(fastq.bam.input, celltag.version,
  extraction.output.filename, save.fullTag = TRUE, save.onlyTag = TRUE)
}
\arguments{
\item{fastq.bam.input}{The input fastq/bam data directory}

\item{extraction.output.filename}{The output file directory, i.e. where would you like to store the 8nt tags with their counts?}

\item{save.fullTag}{Would you like to save full sequences with the short sequences provided without counts? Defaults to TRUE}

\item{save.onlyTag}{Would you like to save only the 8nt CellTags without counts? Defaults to TRUE}

\item{short.nt.before.tag}{A short sequence before the 8nt tag to help more specific identification}

\item{short.nt.after.tag}{A short sequence after the 8nt tag to help more specific identification}
}
\value{
A list contains count table of CellTags. If requested to save fullTag counts, i.e. save.fullTag.counts = TRUE, return a list of both 8nt tags and full sequences count. Otherwise, a list of 8nt tags counts.
}
\description{
This function extracts CellTags from the raw fastq sequencing file, provides counts of each CellTag and sorts them in desending order.
}
\examples{
CellTagExtraction("data.fastq", "v1", "~/Desktop/whitelist.txt")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
