% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagExtraction_Function.R
\name{CellTagMatrixCount}
\alias{CellTagMatrixCount}
\title{CellTag Matrix Generation Function}
\usage{
CellTagMatrixCount(barcodes.file, celltag.read.data.file)
}
\arguments{
\item{barcodes.file}{A .tsv output file from 10x CellRanger pipeline. It contains a list of all cell barcodes identified in the filtered dataset.}

\item{celltag.read.data.file}{A .tsv output file that contains information about the parsed CellTag read data}
}
\value{
Cell Barcodes x CellTag matrix
}
\description{
This function uses the extract information from data processed before and generate a Cell Barcode x CellTag matrix
}
\examples{
("data.fastq", "v1", "~/Desktop/whitelist.txt")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
