% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagWhitelistFiltering_Function.R
\name{CellTagWhitelistFiltering}
\alias{CellTagWhitelistFiltering}
\title{CellTag Whitelist Filtering Function}
\usage{
CellTagWhitelistFiltering(count.sorted.table, percentile, output.dir,
  output.count.file, save.output = TRUE)
}
\arguments{
\item{count.sorted.table}{The count table for all CellTags with first column being CellTag and second being counts.}

\item{percentile}{A fraction cutoff percentile for filtering the CellTags e.g. 0.9 for 90th percentile}

\item{output.dir}{Which directory would you like to store these files?}

\item{output.count.file}{Which CSV filename would you like to save to for the filtered count table? Default to NULL. If NULL but save.output = TRUE, automatic names will be generated and used for storage}

\item{save.output}{Would you like to save your whitelisted counts to file? Default to TRUE}
}
\value{
Whitelisted CellTag count table
}
\description{
This function conducts whitelist filtering such that only CellTags with count number over their certain percentile would be considered for clone calling
}
\examples{
CellTagWhitelistFiltering("data.txt", 0.9, "~/Desktop/", "My_Favourite_Whitelist.csv")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
