% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{CloneCalling}
\alias{CloneCalling}
\title{Clone Calling Function}
\usage{
CloneCalling(Jaccard.Matrix, output.dir, output.filename,
  correlation.cutoff)
}
\arguments{
\item{Jaccard.Matrix}{The Jaccard matrix}

\item{output.dir}{Which directory would you like to save your clone table information?}

\item{output.filename}{Which CSV file would you like to save the clone table into?}

\item{correlation.cutoff}{Correlation cutoff for clone membership}
}
\value{
A list of clone table with membership of each cell and clone size counting table.
}
\description{
This function conducts clone calling based on the Jaccard results
}
\examples{
CloneCalling(Jac, "~/Desktop/", "My_Favourite_CloneCalling.csv", 0.7)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
