% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{JaccardAnalysis}
\alias{JaccardAnalysis}
\title{Jaccard Analysis Function}
\usage{
JaccardAnalysis(whitelisted.celltag.data, save.mtx = TRUE,
  plot.corr = TRUE, save.plot = TRUE)
}
\arguments{
\item{whitelisted.celltag.data}{The single-cell data with whitelisted celltag.}

\item{save.mtx}{Would you like to save the jaccard matrix? Default to TRUE. Save the matrix to the working directory.}

\item{plot.corr}{Would you like to plot the correlation matrix?}

\item{save.plot}{Would you like to save the plot? Save the plot to the working directory.}
}
\value{
The jaccard matrix
}
\description{
This function conducts Jaccard analysis to calculate the Jaccard similarity between cells
}
\examples{
JaccardAnalysis(celltags.whitelisted)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
