% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{MetricBasedFiltering}
\alias{MetricBasedFiltering}
\title{Metric-Base Filtering Function}
\usage{
MetricBasedFiltering(whitelisted.celltag.data, cutoff,
  comparison = "less")
}
\arguments{
\item{whitelisted.celltag.data}{The single-cell data with whitelisted celltag.}

\item{cutoff}{The cutoff decided from the metric plots}

\item{comparison}{Would you like to maintain the part less than/greater than the cutoff? Default to less. Choices can be greater or less.}
}
\value{
A new scRNA-seq data frame with CellTags that passes the metric-based filtering
}
\description{
This function applies further filtering on scRNA-seq data with CellTags based on cutoff values identified from the metric plots
}
\examples{
MetricBasedFiltering(celltags.whitelisted, 20, "less")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
