% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{MetricPlots}
\alias{MetricPlots}
\title{CellTag Metric Plotting Function}
\usage{
MetricPlots(celltag.data)
}
\arguments{
\item{celltag.data}{The single-cell data with celltags}
}
\value{
A list contain the average occurrence of each celltag across cells and the average frequency of different celltags in each cell
}
\description{
This function provides some metric plots for further downstream celltag filtering in the scRNA-seq dataset
}
\examples{
MetricPlots(celltags.whitelisted)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
