% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{SingleCellDataBinarization}
\alias{SingleCellDataBinarization}
\title{Single-cell RNA-seq Whitelisting Function}
\usage{
SingleCellDataBinarization(celltag.dat, tag.cutoff)
}
\arguments{
\item{celltag.dat}{raw celltag single-cell data matrix}

\item{tag.cutoff}{How many tags would you like to be used as a cutoff to say that the cells are tagged?}
}
\value{
Binarized single-cell dataset
}
\description{
This function conducts whitelist filtering through the single-cell dataset
}
\examples{
SingleCellDataWhitelist(sc.mtx, 2)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
