% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagCloneCalling_Function.R
\name{SingleCellDataWhitelist}
\alias{SingleCellDataWhitelist}
\title{Single-cell RNA-seq Whitelisting Function}
\usage{
SingleCellDataWhitelist(celltag.dat, whitels.cell.tag.file)
}
\arguments{
\item{celltag.dat}{binarized celltag single-cell data matrix}

\item{whitels.cell.tag.file}{file director to the whitelisted cell tags}
}
\value{
Whitelisted single-cell dataset
}
\description{
This function conducts whitelist filtering through the single-cell dataset
}
\examples{
SingleCellDataWhitelist(sc.mtx, "~/Desktop/My_Favourite_Whitelist.csv", 2)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
